refactor(noodle): split public generation services
